Fix full ADDER and add 4-input OR
The bangs were routed incorrectly and hence were not forcing real-time
hot update of overall system, exhibiting cached states.

The new pattern uses a bang to "hot-push" from the most-right unit and
bubble to the left until trigger the hot inlet of the bottom-most OR gate.
Also messages representing a ruth table for a one-bit binary adder (three columns)
are added.